Modify Landsat URL generation to fall through data sources for L1GT
This issue was due to the Landsat data type "L1GT" having an ambiguous
meaning as far as whether the scene is a Collection-1 scene or not. I
am changing the logic for obtaining Collection-1 to account for this
ambiguity by using the following logic instead:

1. if this is neither a Collection-1 scene nor a pre-Collection scene,
   return an error for unknown data type
2. if this is a Collection-1 scene, then try to look it up in the scene
   list index downloaded from AWS; if it exists there, return the URL
   specified there; if not, continue to next step
3. if this is a pre-Collection scene, then return the URL using our
   old method of parsing the scene ID fail with a "not found" error
4. The pre-Collection types are ["L1T", "L1GT", "L1G"], while the
   Collection-1 types are ["L1TP", "L1GT", "L1GS"]. This gives the
   necessary "fallthrough" behavior to correctly resolve L1GT assets.

Unit tests are included to verify this behavior.